flattened packages
